Added the finding peakvalue element code
Problem statement: Peak Element in an Array • The peak element in an array is an array element which is not smaller than it's  neighbours.  
• For example, given an array of {6,7,10,12,9} 12 is the peak element of the array.  
• Another example is an array of {8,15,9,2,23,5} in this case, there are two peak  elements:15 and 23. 
However, there are some edge cases to be considered. 
• One edge case can be when all elements in the array are the same such as {5,5,5}.  The peak element is 5. 
• The second edge case to consider is when the elements in the array are sorted in  descending order i.e., from the highest to the lowest.{50,40,30,20}.50 is the peak  element. 
• The final edge is elements in an array sorted in ascending order i.e., the lowest  element to the highest element {20,30,40,50,70,90}. The peak element is 90. 

The peak element of an array can be found using the naive approach of linear  search with time complexity O(N) or the optimized divide and conquer approach  with time complexity O(logn).

The Linear Search Approach 

How it works & Implementation 
In this approach we traverse through the array and compare elements with it's  neighbours. Once a peak element is found, it's immediately returned. 
procedure linear_search (list, value) 
 for each item in the list 
 if match item == value 
 return the item's location 
 end if 
 end for 
end procedure



Steps for Solving: 
Linear Search (Array A, Value x) 
Step 1: Set i to 1 
Step 2: if i > n then go to step 7 
Step 3: if A[i] = x then go to step 6 
Step 4: Set i to i + 1 
Step 5: Go to Step 2 
Step 6: Print Element x Found at index i and go to step 8 Step 7: Print element not found 
Step 8: Exit